Security vulnerability fixes - CSRF token - 4th nov 2022

+++ includes/class-wcj-payment-gateways.php
@@ -47,10 +47,11 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 		 * @since   3.3.0
 		 */
 		public function maybe_delete_payment_gateway_input_fields() {
-			if ( isset( $_GET['wcj_delete_payment_gateway_input_fields'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification
-				$order_id = sanitize_text_field( wp_unslash( $_GET['wcj_delete_payment_gateway_input_fields'] ) ); // phpcs:ignore WordPress.Security.NonceVerification
+			$wpnonce = isset( $_GET['_wpnonce'] ) ? wp_verify_nonce( sanitize_key( $_GET['_wpnonce'] ), 'wcj_delete_payment_gateway_input_fields' ) : false;
+			if ( $wpnonce && isset( $_GET['wcj_delete_payment_gateway_input_fields'] ) ) {
+				$order_id = sanitize_text_field( wp_unslash( $_GET['wcj_delete_payment_gateway_input_fields'] ) );
 				delete_post_meta( $order_id, '_wcj_custom_payment_gateway_input_fields' );
-				wp_safe_redirect( esc_url( remove_query_arg( 'wcj_delete_payment_gateway_input_fields' ) ) );
+				wp_safe_redirect( esc_url_raw( remove_query_arg( array( 'wcj_delete_payment_gateway_input_fields', '_wpnonce' ) ) ) );
 				exit;
 			}
 		}
@@ -130,7 +131,7 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 				)
 			);
 			if ( 'yes' === wcj_get_option( 'wcj_custom_payment_gateways_input_fields_delete_button', 'no' ) ) {
-				$html .= '<p><a style="color:#a00;" href="' . esc_url( add_query_arg( 'wcj_delete_payment_gateway_input_fields', $order_id ) ) . '"' . wcj_get_js_confirmation() . '>' .
+				$html .= '<p><a style="color:#a00;" href="' . esc_url( wp_nonce_url( add_query_arg( 'wcj_delete_payment_gateway_input_fields', $order_id ), 'wcj_delete_payment_gateway_input_fields' ) ) . '"' . wcj_get_js_confirmation() . '>' .
 				__( 'Delete', 'woocommerce-jetpack' ) . '</a></p>';
 			}
 			echo wp_kses_post( $html );
